+++ Changelog.md
@@ -4,6 +4,9 @@ Only versions past 2.1 have their changes documented here.
 
 Source code and usage examples can be found on Github.
 
+### v2.2-dev4 [Nov 08, 2019]
+* Added `DMKText` object and general use `CreateText()` function for text object creation 
+
 ### v2.2-dev3 [Nov 03, 2019]
 * Moved existing A1 variants to the A2 designation and created separate A1 designations for all existing bullet creation functions  
 * Implemented `CreateShotA1()` and `CreateShotA2()`